Move SlackClient to a package
